Combine config files
