--- api/clients/v2/payload_retriever_config.go
@@ -1,12 +1,0 @@
-package clients
-
-import "time"
-
-// PayloadRetrieverConfig contains an embedded PayloadClientConfig, plus all additional configuration values needed
-// by a PayloadRetriever
-type PayloadRetrieverConfig struct {
-	PayloadClientConfig
-
-	// The timeout duration for relay calls to retrieve blobs.
-	RelayTimeout time.Duration
-}
